Fixed game directory issue.
Had an issue where the game directory was incorrectly auto created. 

If you have issues opening the game, make sure the game directory is set in the config.ini file. 
It should look like "[Game Directory]=C:\ProgramData\KingsIsle Entertainment\Wizard101\Bin\WizardGraphicalClient.exe", and not "[Game Directory]=:"
You set this in the settings in the program, and not manually. MAKE SURE TO HIT SAVE IN SETTINGS!